Fix clang narrowing errors in chat-log brace initializers
xd/yd are ints; brace-initializing a doublePair from them is an
int-to-double narrowing conversion. GCC accepts it (at most a
-Wnarrowing diagnostic) but clang rejects it outright:

  error: non-constant-expression cannot be narrowed from type 'int'
  to 'double' in initializer list [-Wc++11-narrowing]

Cast explicitly; the values are exact in double, so behavior is
unchanged. This unbreaks any clang build, not just macOS.

+++ gameSource/yummyLife.cpp
@@ -157,7 +157,7 @@ void YummyLife::ChatLog::refreshDistancesForCurrentPosition() {
     LiveObject *us = HetuwMod::ourLiveObject;
     if (us == NULL) return;
 
-    doublePair currentPos = { us->xd, us->yd };
+    doublePair currentPos = { (double)us->xd, (double)us->yd };
     if (lastDistanceCalcPos.x == currentPos.x &&
         lastDistanceCalcPos.y == currentPos.y) {
         return;
